requirements.txt
# ============================================================
# DEPENDENCIAS DEL PROYECTO
# Sistema de Reconocimiento Facial con Supabase
# ============================================================

# Visión por computadora
opencv-python==4.10.0.84
opencv-contrib-python==4.10.0.84
imutils==0.5.4

# Procesamiento de datos
numpy==1.26.4

# Base de datos
supabase==2.9.0
python-dotenv==1.0.1

# Interfaz gráfica (incluida en Python estándar)
# tkinter (viene con Python)

# ============================================================
# INSTALACIÓN:
# pip install -r requirements.txt
# ============================================================